// src/services/api/models/AliyunOcrRecognizeAllTextType.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum AliyunOcrRecognizeAllTextType {
    ADVANCED = 'Advanced',
    GENERAL = 'General',
    COMMERCE = 'Commerce',
    HAND_WRITING = 'HandWriting',
    MULTI_LANG = 'MultiLang',
    TABLE = 'Table',
    ID_CARD = 'IdCard',
    BANK_CARD = 'BankCard',
    INTERNATIONAL_PASSPORT = 'InternationalPassport',
    CHINESE_PASSPORT = 'ChinesePassport',
    SOCIAL_SECURITY_CARD = 'SocialSecurityCard',
    PERMIT_TO_HK_MO_TW = 'PermitToHK_MO_TW',
    PERMIT_TO_MAINLAND = 'PermitToMainland',
    HOUSEHOLD_HEAD = 'HouseholdHead',
    HOUSEHOLD_RESIDENT = 'HouseholdResident',
    ESTATE_CERTIFICATION = 'EstateCertification',
    BIRTH_CERTIFICATION = 'BirthCertification',
    HKID_CARD = 'HKIdCard',
    INTERNATIONAL_ID_CARD = 'InternationalIdCard',
    STAMP = 'Stamp',
    MIXED_INVOICE = 'MixedInvoice',
    INVOICE = 'Invoice',
    CAR_INVOICE = 'CarInvoice',
    QUOTA_INVOICE = 'QuotaInvoice',
    AIR_ITINERARY = 'AirItinerary',
    TRAIN_TICKET = 'TrainTicket',
    TOLL_INVOICE = 'TollInvoice',
    ROLL_TICKET = 'RollTicket',
    BANK_ACCEPTANCE = 'BankAcceptance',
    BUS_SHIP_TICKET = 'BusShipTicket',
    NON_TAX_INVOICE = 'NonTaxInvoice',
    COMMON_PRINTED_INVOICE = 'CommonPrintedInvoice',
    HOTEL_CONSUME = 'HotelConsume',
    PAYMENT_RECORD = 'PaymentRecord',
    PURCHASE_RECORD = 'PurchaseRecord',
    RIDE_HAILING_ITINERARY = 'RideHailingItinerary',
    SHOPPING_RECEIPT = 'ShoppingReceipt',
    TAX_CLEARANCE_CERTIFICATE = 'TaxClearanceCertificate',
    USED_CAR_INVOICE = 'UsedCarInvoice',
    VEHICLE_LICENSE = 'VehicleLicense',
    DRIVING_LICENSE = 'DrivingLicense',
    VEHICLE_REGISTRATION = 'VehicleRegistration',
    VEHICLE_CERTIFICATION = 'VehicleCertification',
    LICENSE_PLATE_NUMBER = 'LicensePlateNumber',
    CAR_VIN_CODE = 'CarVinCode',
    BUSINESS_LICENSE = 'BusinessLicense',
    INTERNATIONAL_BUSINESS_LICENSE = 'InternationalBusinessLicense',
    MEDICAL_DEVICE_MANAGE_LICENSE = 'MedicalDeviceManageLicense',
    MEDICAL_DEVICE_PRODUCE_LICENSE = 'MedicalDeviceProduceLicense',
    COSMETIC_PRODUCE_LICENSE = 'CosmeticProduceLicense',
    QR_CODE = 'QrCode',
    BAR_CODE = 'BarCode',
    TAXI_INVOICE = 'TaxiInvoice',
    TRADEMARK_CERTIFICATE = 'TrademarkCertificate',
    FOOD_PRODUCE_LICENSE = 'FoodProduceLicense',
    FOOD_MANAGEMENT_LICENSE = 'FoodManagementLicense',
    CLASS_IIMEDICAL_DEVICE_MANAGE_LICENSE = 'ClassIIMedicalDeviceManageLicense',
    WAY_BILL = 'WayBill',
    BANK_ACCOUNT_PERMIT = 'BankAccountPermit',
}